+++ PROYECTO_ENCOMIENDA/src/proyecto_encomienda/LOGISTICA_TRANSPORTE/Main_Transporte.java
@@ -5,7 +5,7 @@ import  proyecto_encomienda.LOGISTICA_TRANSPORTE.Camion;
 import  proyecto_encomienda.LOGISTICA_TRANSPORTE.Ubicacion;
 import  proyecto_encomienda.LOGISTICA_TRANSPORTE.Vehiculo;
 import proyecto_encomienda.GESTION_PAQUETES.BACKEND.Paquete;
-import GESTOR_PERFILES.Conductor;
+import proyecto_encomienda.GESTOR_PERFILES.Conductor;
 
 import java.util.ArrayList;
 import java.util.List;
