+++ PROYECTO_ENCOMIENDA/src/proyecto_encomienda/PROYECTO_ENCOMIENDA.java
@@ -1,7 +1,7 @@
 
 package proyecto_encomienda;
 
-import GESTOR_PERFILES.GUIs.TipoAcciónGUI;
+import proyecto_encomienda.GESTOR_PERFILES.GUIs.TipoAcciónGUI;
 import proyecto_encomienda.GUI.JFIngresar;
 
 public class PROYECTO_ENCOMIENDA {
